requirements.txt
requests>=2.32.0
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
beautifulsoup4>=4.12.0
pypdf>=5.0.0
python-docx>=1.1.0
httpx>=0.28.0
pytest>=9.0.0
playwright>=1.52.0
